+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/impl/DynamicOrphanedSnapshotsDetector.java
@@ -11,6 +11,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import de.skuzzle.test.snapshots.SnapshotTestResult;
+import de.skuzzle.test.snapshots.impl.OrphanDetectionResult.Result;
 import de.skuzzle.test.snapshots.io.UncheckedIO;
 import de.skuzzle.test.snapshots.validation.Arguments;
 
@@ -36,25 +37,31 @@ final class DynamicOrphanedSnapshotsDetector {
         this.failedTestMethods.add(Arguments.requireNonNull(testMethod));
     }
 
-    public Stream<Path> detectOrphans(Path snapshotDirectory) {
+    public Stream<OrphanDetectionResult> detectOrphans(Path snapshotDirectory) {
         try (final var files = UncheckedIO.list(snapshotDirectory)) {
             return files
-                    .filter(this::isOrphanedSnapshot)
+                    .filter(this::isSnapshotFile)
+                    .map(this::isOrphanedSnapshot)
                     .collect(Collectors.toList())
                     .stream();
         }
     }
 
-    private boolean isOrphanedSnapshot(Path potentialSnapshotFile) {
-        if (!InternalSnapshotNaming.isSnapshotFile(potentialSnapshotFile)) {
-            return false;
-        }
+    private boolean isSnapshotFile(Path potentialSnapshotFile) {
+        return InternalSnapshotNaming.isSnapshotFile(potentialSnapshotFile);
+    }
+
+    private OrphanDetectionResult isOrphanedSnapshot(Path snapshotFile) {
         // we can not detect orphaned snapshots for failed tests because the test might
         // have failed before creating the snapshot
-        if (pertainsToFailedTest(potentialSnapshotFile)) {
-            return false;
+        Result result;
+        if (pertainsToFailedTest(snapshotFile)) {
+            result = Result.UNSURE;
+        } else {
+            final boolean containedInTest = testResultsContain(snapshotFile);
+            result = containedInTest ? Result.ACTIVE : Result.ORPHAN;
         }
-        return !testResultsContain(potentialSnapshotFile);
+        return new OrphanDetectionResult(getClass().getSimpleName(), snapshotFile, result);
     }
 
     private boolean pertainsToFailedTest(Path snapshotFile) {
@@ -69,4 +76,5 @@ final class DynamicOrphanedSnapshotsDetector {
                 .anyMatch(snapshotFileFromResult -> UncheckedIO.isSameFile(snapshotFileFromResult,
                         snapshotFile));
     }
+
 }