+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/impl/SnapshotTestContext.java
@@ -2,7 +2,9 @@ package de.skuzzle.test.snapshots.impl;
 
 import java.lang.reflect.Method;
 import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -11,6 +13,7 @@ import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
 
 import de.skuzzle.test.snapshots.EnableSnapshotTests;
+import de.skuzzle.test.snapshots.SnapshotDsl.ChooseDirectory;
 import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
 import de.skuzzle.test.snapshots.SnapshotTestResult;
 import de.skuzzle.test.snapshots.impl.SnapshotLogging.SnapshotLogger;
@@ -35,6 +38,8 @@ public final class SnapshotTestContext {
 
     private final DynamicOrphanedSnapshotsDetector dynamicOrphanedSnapshotsDetector = new DynamicOrphanedSnapshotsDetector();
     private final SnapshotConfiguration snapshotConfiguration;
+
+    private final List<Path> dynamicSnapshotDirectories = new ArrayList<>();
 
     private SnapshotTestImpl currentSnapshotTest;
 
@@ -114,6 +119,17 @@ public final class SnapshotTestContext {
     }
 
     /**
+     * Records an extra directory for containing snapshots. The framework uses this to
+     * discover dynamic snapshot directories that are use with
+     * {@link ChooseDirectory#in(Path)}.
+     *
+     * @param dynamicSnapshotDirectory The snapshot directory to add.
+     */
+    public void recordDynamicSnapshotDirectory(Path dynamicSnapshotDirectory) {
+        dynamicSnapshotDirectories.add(dynamicSnapshotDirectory);
+    }
+
+    /**
      * Records the results from all snapshot assertions within a single test method.
      *
      * @param results The snapshot test results of a single test method.
@@ -131,10 +147,15 @@ public final class SnapshotTestContext {
     public Collection<Path> detectOrCleanupOrphanedSnapshots() {
         final boolean deleteOrphaned = this.snapshotConfiguration.isDeleteOrphanedSnapshots();
 
-        final Path snapshotDirectory = snapshotConfiguration.determineSnapshotDirectory();
-        final Stream<Path> dynamicOrphans = dynamicOrphanedSnapshotsDetector.detectOrphans(snapshotDirectory);
-        final Stream<Path> staticOrphans = new StaticOrphanedSnapshotDetector().detectOrphans(DirectoryResolver.BASE);
-        return Stream.concat(dynamicOrphans, staticOrphans)
+        final Stream<OrphanDetectionResult> dynamicOrphans = allSnapshotDirectories()
+                .flatMap(dynamicOrphanedSnapshotsDetector::detectOrphans);
+
+        final Stream<OrphanDetectionResult> staticOrphans = new StaticOrphanedSnapshotDetector()
+                .detectOrphans(DirectoryResolver.BASE);
+
+        return new OrphanPostProcessor()
+                .orphanedOnly(Stream.concat(dynamicOrphans, staticOrphans).collect(Collectors.toList()))
+                .map(OrphanDetectionResult::snapshotFile)
                 .distinct()
                 .peek(orphaned -> {
                     if (deleteOrphaned) {
@@ -149,6 +170,11 @@ public final class SnapshotTestContext {
                     }
                 })
                 .collect(Collectors.toList());
+    }
+
+    private Stream<Path> allSnapshotDirectories() {
+        final Path defaultSnapshotDirectory = this.snapshotConfiguration.determineSnapshotDirectory();
+        return Stream.concat(Stream.of(defaultSnapshotDirectory), dynamicSnapshotDirectories.stream());
     }
 
 }