+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/impl/SnapshotTestImpl.java
@@ -12,6 +12,7 @@ import org.opentest4j.AssertionFailedError;
 
 import de.skuzzle.test.snapshots.SnapshotDsl.ChooseActual;
 import de.skuzzle.test.snapshots.SnapshotDsl.ChooseDataFormat;
+import de.skuzzle.test.snapshots.SnapshotDsl.ChooseName;
 import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
 import de.skuzzle.test.snapshots.SnapshotException;
 import de.skuzzle.test.snapshots.SnapshotFile;
@@ -34,7 +35,9 @@ final class SnapshotTestImpl implements Snapshot, InternalSnapshotTest {
     private final Method testMethod;
     private final SnapshotConfiguration configuration;
     private final LocalResultCollector localResultCollector = new LocalResultCollector();
+
     private SnapshotNaming namingStrategy = SnapshotNaming.defaultNaming();
+    private Path directoryOverride;
 
     SnapshotTestImpl(SnapshotConfiguration configuration, Method testMethod) {
         this.configuration = Arguments.requireNonNull(configuration, "configuration must not be null");
@@ -44,6 +47,12 @@ final class SnapshotTestImpl implements Snapshot, InternalSnapshotTest {
     @Override
     public ChooseActual namedAccordingTo(SnapshotNaming namingStrategy) {
         this.namingStrategy = Arguments.requireNonNull(namingStrategy, "namingStrategy must not be null");
+        return this;
+    }
+
+    @Override
+    public ChooseName in(Path directory) {
+        this.directoryOverride = Arguments.requireNonNull(directory, "snapshot directory must not be null");
         return this;
     }
 
@@ -62,7 +71,16 @@ final class SnapshotTestImpl implements Snapshot, InternalSnapshotTest {
 
     private Path determineSnapshotFile(String snapshotName) throws IOException {
         final String snapshotFileName = InternalSnapshotNaming.getSnapshotFileName(snapshotName);
-        return configuration.determineSnapshotDirectory().resolve(snapshotFileName);
+        return determineSnapshotDirectory().resolve(snapshotFileName);
+    }
+
+    private Path determineSnapshotDirectory() {
+        return this.directoryOverride != null ? this.directoryOverride : configuration.determineSnapshotDirectory();
+    }
+
+    @Override
+    public Optional<Path> snapshotDirectoryOverride() {
+        return Optional.ofNullable(directoryOverride);
     }
 
     SnapshotTestResult justUpdateSnapshotWith(SnapshotSerializer snapshotSerializer, Object actual) throws Exception {