+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/junit5/JUnit5SnapshotExtension.java
@@ -59,6 +59,7 @@ public final class JUnit5SnapshotExtension implements
 
         final InternalSnapshotTest snapshotTest = snapshotTestContext.clearCurrentSnapshotTest().orElse(null);
         if (snapshotTest != null) {
+            snapshotTest.snapshotDirectoryOverride().ifPresent(snapshotTestContext::recordDynamicSnapshotDirectory);
             snapshotTestContext.recordSnapshotTestResults(snapshotTest.testResults());
             snapshotTest.executeAssertions();
         }